backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
aiofiles==23.2.1
python-multipart==0.0.6
Pillow==10.1.0
gradio_client==0.7.0
requests==2.31.0

